refactor(spa): rename routes/ directory to views/
The SPA no longer uses URL-based routing after ADR-0011 removed the
router. The 'routes' directory now holds view renderers, so this
renames it to 'views' for accuracy.

Changes:
- Renamed src/spa/routes/ → src/spa/views/ via git mv
- Updated all import paths in source and test files
- Updated comment-only references in source code, e2e specs, and docs

No behavioral change — imports are updated atomically with the move.
