+++ src/spa/views/game.ts
